requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
openai==1.40.0
pydantic==2.9.2
python-dotenv==1.0.1
